Converted all the Share activities to menu activity and removed classes for custom activity type lists
