--- Sources/Brave/Frontend/Share/CopyCleanURLActivity.swift
@@ -1,39 +1,0 @@
-// Copyright 2021 The Brave Authors. All rights reserved.
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-import Foundation
-import Shared
-import UIKit
-import SwiftUI
-
-/// An activity that will copy clean URL of the webpage
-class CopyCleanURLActivity: UIActivity, MenuActivity {
-  fileprivate let callback: () -> Void
-
-  init(callback: @escaping () -> Void) {
-    self.callback = callback
-  }
-
-  override var activityTitle: String? {
-    return Strings.copyCleanLink
-  }
-
-  override var activityImage: UIImage? {
-    UIImage(braveSystemNamed: "leo.broom")?.applyingSymbolConfiguration(.init(scale: .large))
-  }
-  
-  var menuImage: Image {
-    Image(braveSystemName: "leo.broom")
-  }
-
-  override func perform() {
-    callback()
-    activityDidFinish(true)
-  }
-
-  override func canPerform(withActivityItems activityItems: [Any]) -> Bool {
-    return true
-  }
-}
